NHSO-334 - show header and navbar when logged in
Moved existing Nav bar from Home component to App.


Added new Header menu, also in App.
Wired up Nav bar More button click event to router to more view, in order to test that header and nav menus are shown from other views, not just Home.
Header and Nav bar menus to be shown if and only if the user is logged in and the user is not using a native app.