Bug with environment variables
There was a bug caused by the introduction of an environment variable in
the `npm run` script that did not work on Windows.  Discovered that the
environment variable approch has been superceded so refactoring the
`LoginButton.vue` and authorisation service to use the `$config`
property instead.

Refactored webpack to remove the vestigial environment configurations