implemented view my booked appointments scenarios
Updated xpaths for server error; added page check for progressing to appointment booking page; updated "My appointments" title; removed bug tag.

Reverted check addition as it caused a problem with another test, but doesn't seem to be an issue anymore anyway.

Removed the timezone setting when creating Emis data (should always be local).

Started to refactor the AppointmentData to improve NHSO-196, but have set it to pending so we can sign off NHSO-524.