BDD for Cancel Appointments page (excluding Cancel button)
* Updates to some of the locators (tech-debt task raised to continue this);
* Scenarios implemented for NHSO-663;
* Some refactoring for Appointments.

Another tweak to the upcoming appointments locator.

Fixed failing tests for symptoms.

Moving the new files back to where the should be!

Unblock Appointments Feature files.

Added a new step for when we progress to the Appointment Booking page, but we don't necessarily expect it to load before the error page.

Added step back in that previously allowed this test to work, but added in the next Step definition.