Initial commit.
1st iteration of implementing the scenario.

Update following review comments on deleted MR+branch for BDD skeleton for NHSO-2079.

A bit of refactoring of AppointmentsSlotsExample, but some can be reverted.

Altered the way we Store Date and Time of the Slot we are going to select. Will need improving and haven't tested it yet.

Fixed issue with the storing of the slot to be selected.

Almost finished implementing the "Tomorrow" test. "Today" test already exists.

Created another Step Definition and also removed duplicated Scenarios that must have occurred via rebasing.

Have the "all available" test almost working, just need an extra assertion (currently marked as a TODO).

Completed most of the "This Week" test, including some improvement in checking for the expected number of slots. Still need to create methods to determine the expected Dates for This Week and Next Week.

Should have complete for This Week and not much to do for Next Week. Got some debugging in after removing suppressions.

Finished implementing the steps, but can't yet test due to environment issues.

Confirmed that tests are now working. Had to revert page properties back to functions as it makes "the page look dodgy" so had to add suppression back in. Have removed the "too generic exception" suppressions and used "NoSuchElementException".

Fixed introduced test failures and resolve own discussion.

Added fix for a flaky test. Went with the waitForSpinnerToDisappear, as opposed to a iWait(seconds = 10) (which is explicit in the BDD Step).

Attempt at tidying xpaths in availableAppointmentsPage. Only tested one Scenario so far.

Refactored to use new page element assert methods.

Further changes for review. Not tested yet, though.

Updated DateTimeWrapper as properties weren't being initiliased in the correct order. Also, updated the hours and minutes parameters as completely optional as they are only used to override the localDateTime. For this reason, localDateTime should only now be updated with the presence of either of these arguments.

Fixes for some failing tests.

More changes for review.

Added helpful names for the HybridPageElements on AvailableAppointmentsPage.