A step needs reimplementing. Will resolve as part of this branch.
Had previously noticed a commented line with a bug reference that has since been closed. Reintroduced step, but there's a new defect, so have blocked those test.

Completed updating tests so that they no longer use "that" method and removed "that" method. Using the ErrorPage instead.

Got rid of the need for WEBDRIVER_BASE_URL because "it was annoying me". Hopefully, there's actual need for it... Also, there was some broken logic.

Moved the assertions into a single ErrorPage method. For whatever reason, title and messageText isn't always being passed, so will only check when they are not null. Hopefully, these values can be added when the Error Messages are updated.

A bit more tidy-up. Removed AuthReturnSteps and moved contents to either the PageObject or StepDefinitions.

Fixed some broken tests.