Azure Cosmos DB DocumentClient instances should be singletons
As per https://docs.microsoft.com/en-us/azure/cosmos-db/performance-tips

"Use a singleton Azure Cosmos DB client for the lifetime of your application

Each DocumentClient instance is thread-safe and performs efficient connection management and address caching when operating in Direct Mode. To allow efficient connection management and better performance by DocumentClient, it is recommended to use a single instance of DocumentClient per AppDomain for the lifetime of the application."

To achieve this, TermsAndConditionsService should be registered with the DI container with Singleton lifestyle, not Scoped.