NHSO-2675 -  A bit of refactoring is needed prior to introducing Vision. Will establish separate branch.
Made a correction in the backend worker readme.

Made a start on the refactoring. Only really started populating the MyAppointmentsFacade. The fields may be unnecessary.

Have refactored for no upcoming appointments and verified it works for both Emis and Tpp.

Got scenarios working for both suppliers for upcoming appointments, including multiple, at the same time.

Opted towards using the generic example, rather than the single slot. Reintroduced the ability to have multiple clinicians for a session and have currently left 1 of the sessions in the generic example to have 2.

Refactored to use mappings and replaced the commented code. Still need to verify the changes have worked, though.

Changes as per review comments and to resolve failing tests. I still want to remove the "SLOT_ID" constant as it's rubbish.

Refactored to reuse the storage of expected slots when cancelling the first appointment. Added id to Slot. Even though it's not displayed on the UI, it's obviously required by the UI.

Removed the additional clinician on generic example. Future task will create another example for use with a Scenario for EMIS.

Changes to meet Detekt standards.

Review changes.

Reverted inadvertant change that broke the clinician filter.

Fixed issue where the Emis stubs for Appointment Slots were not representative of the Emis response according to their Swagger contracts as we were passing the facade in directly.

Fixed a failing test and resolved more review comments.

Used distinct() instead of converting from list to set and back again.

Pushed mockUpcomingAppointments to base class.

Changed Exception to NotImplementedError.