NHSO-2463
Refactored the AppointmentsSlotsExampleBuilder files. Renamed the one with expectations and made it the child of the other.

Removed duplicate AppointmentSessionFacadeBuilder as no real difference. A bit of tidy of the stubs/appointments package.

Moved the content of stubs/appointments/AppointmentsSlotsExample used by AppointmentSlotsStubs to that class, as that's only place they'll be used.

Re-resolved Detekt issues reintroduced via recent changes to Develop and rebasing.

Removed wildcard imports. Reran detekt - no issues found.

Tidied imports for AppointmentsSlotsFactory.