NHSO-3074 Vision appointments handle access denied response
Need to handle the situation where Vision appointment API call return Access Denied (Vision error code -35).

Although we are already performing a check on VisionUserSession.IsAppointmentsEnabled immediately prior to making the Vision API request, there is a possibility that this is out dated, if the practice has changed configuration since the user session was created.