NHSO-3086 correct redundant exception handling, fix tests
BDD tests being developed for NHSO-800 revealed that VisionAppointmentSlotsService was incorrectly catching generic System.Exception. This prevented timeouts in Vision API calls from correctly being handled by the TimeoutExceptionFilterAttribute, and ultimately returning a 504 Gateway Timeout.

In fixing this, identified that one of the unit tests was not testing the thing its title described.