NHSO-3277: Entry/exit logging consistency
Whilst we have appropriate entry/exit logging, it has been noted that our approach to it is inconsistent across the codebase. Some places call a custom extension
method, some call a raw log method passing in the name of the method being entered.
Further, there is a feature in C# that allows you to get the name of the calling method at compile-time without explicitly passing it in:
https://docs.microsoft.com/en-us/dotnet/api/system.runtime.compilerservices.callermembernameattribute
As this is a compile-time feature it suffers no runtime performance penalty.

The scope of this ticket is to:

Change the extension method to make use of the CallerMemberName attribute
Identify all enter/exit log entries that don't use the extension method and change them to use it|