nhso-5040 - android refactor around handling urls
webview.url value was causing an error in android 9 coming from an external page to an nhs app page. Known services added query params and reloaded the view, but the webview.url still had the previous url in it, causing it to get into a loop.

Also making the links on the first page behave like spa navigation instead of reloading the whole app.