NHSO-7311: Appointment Slot Type Transformer
Appointment slot types of "Default" will be replaced with "General".

Modified  appointment slot mappers to return IList<Slot> instead of IEnumerable<Slot>.
The latter was leading to the enumerable being evaluated four times - when transforming, logging, auditing, and finally outputting the slots.
Evaluating the results of the IEnumerable to an in-memory list once has an overall performance benefit.