NHSO-9061: tweak performance-affecting config
Allow the users/info/messages pods to access the minimock pods for perf
tests

Remove memory-based scaling as .net and node services will consume all
they can and not release it. We should allow as much memory as is needed
and scale based on CPU usage.

Add a role to the wiremock pod so that it appears in the relevant Splunk
dashboards and allows easy viewing of resource usage.