NHSO-10887: ensure each nominated pharmacy update request uses the latest config singleton
When the application starts, it tries to contact Spine to retrieve some details. If it's successful it updates the config (which is a singleton). If this service is instantiated before the config is updated, it previously would've always had the default empty value for path. This change ensures that each request uses the latest path value from the config.