NHSO-10884: Fixing iOS issue where selecting try again after internet disconnected results in gray screen
Pre-nuxt removal the url passed to the native apps were prefixed with "/". iOS is always removing the first character of this string. Since nuxt removal the URLs passed to the native app are without the "/" prefix so it was removing the first letter of the URL to retry after internet was reconnected e.g. it would try to go to "<base url>/rescriptions" instead of "<base url>/prescriptions".