NHSO-13005: linked accounts refactor part 2
Maintaining a mapping of temporary patient session ids (for the main patient and their proxies) to their constant patient ids from their gp system (PatientActivityContextGuid for Emis, PatientId for TPP). This allows for a gp session to be recreated but the ids to remain the same (for the duration of the session).