NHSO-22252: Check back function exists before calling back() on third party sites
A third party has seen more "script error." errors since release 2.13 where we released back button functionality for third parties. Adding a guard around the back button call so if the user presses the back button but the third party page does not contain (or has not loaded) our nhsapp.js file that a javascript error is not triggered.