Update nhs-app.yaml
The wording in the Overview was ambiguous. It wasn't clear to the reader whether the API was "a simple and secure way for patients registered with a GP surgery in England to access a range of services on their smartphone or tablet." or whether it was the NHS App. (I know it's the latter)